Add nonclobbering packages that only register if no other compressor has been registered with the same name
These new "nonclobbering" packages are suitable for other reusable
libraries to import, if they do not want to accidentally clobber
their user's previously registered compressors.
