+++ internal/s2/s2_test.go
@@ -1,22 +1,18 @@
-/*
- *
- * Copyright 2021 gRPC authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
+// Copyright 2022 Mostyn Bramley-Moore.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 
-package zstd
+package s2
 
 import (
 	"bytes"
@@ -25,10 +21,8 @@ import (
 	"net"
 	"testing"
 
-	"github.com/klauspost/compress/zstd"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials/insecure"
 	"google.golang.org/grpc/encoding"
@@ -39,35 +33,37 @@ import (
 
 const (
 	bufSize = 1024
-	message = "Message Request zSTD"
+	message = "Message Request s2"
 )
 
 func TestRegisteredCompression(t *testing.T) {
-	for _, lvl := range []zstd.EncoderLevel{zstd.SpeedFastest, zstd.SpeedDefault, zstd.SpeedBetterCompression, zstd.SpeedBestCompression} {
-		require.NoError(t, SetLevel(lvl))
+	clobbering := true
+	PretendInit(clobbering)
 
-		comp := encoding.GetCompressor(Name)
-		require.NotNil(t, comp)
-		assert.Equal(t, Name, comp.Name())
+	comp := encoding.GetCompressor(Name)
+	require.NotNil(t, comp)
+	assert.Equal(t, Name, comp.Name())
 
-		buf := bytes.NewBuffer(make([]byte, 0, bufSize))
-		wc, err := comp.Compress(buf)
-		require.NoError(t, err)
+	buf := bytes.NewBuffer(make([]byte, 0, bufSize))
+	wc, err := comp.Compress(buf)
+	require.NoError(t, err)
 
-		_, err = wc.Write([]byte(message))
-		require.NoError(t, err)
-		assert.NoError(t, wc.Close())
+	_, err = wc.Write([]byte(message))
+	require.NoError(t, err)
+	assert.NoError(t, wc.Close())
 
-		r, err := comp.Decompress(buf)
-		require.NoError(t, err)
-		expected, err := ioutil.ReadAll(r)
-		require.NoError(t, err)
+	r, err := comp.Decompress(buf)
+	require.NoError(t, err)
+	expected, err := ioutil.ReadAll(r)
+	require.NoError(t, err)
 
-		assert.Equal(t, message, string(expected))
-	}
+	assert.Equal(t, message, string(expected))
 }
 
 func TestRoundTrip(t *testing.T) {
+	clobbering := true
+	PretendInit(clobbering)
+
 	lis := bufconn.Listen(bufSize)
 	t.Cleanup(func() {
 		assert.NoError(t, lis.Close())
